requirements.txt
numpy>=1.21.0
tqdm>=4.64.0
tensorboard>=2.10.0
loguru>=0.7.0
pyarrow>=10.0.0
pandas>=1.3.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
jupyter>=1.0.0
jupyterlab>=3.0.0
matplotlib>=3.0.0
rich>=12.0.0
marimo
altair
tiktoken>=0.5.0
